move pearl metrics to bob-common
