+++ bob-backend/src/pearl/mod.rs
@@ -2,11 +2,10 @@ mod core;
 mod data;
 mod group;
 mod holder;
-mod metrics;
 mod settings;
 mod stuff;
 
 #[cfg(test)]
 mod tests;
 
-pub use self::{core::Pearl, group::Group, holder::Holder, metrics::init_pearl};
+pub use self::{core::Pearl, group::Group, holder::Holder};
