+++ bob/src/cleaner.rs
