+++ bob/src/counter.rs
