+++ bob/src/link_manager.rs
